requirements.txt
python-vlc>=3.0
mutagen>=1.47
Pillow>=10.0
tinytag>=2.0
